# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Second-Hand Store API",
    "description": "Fully autonomous second-hand store with AI negotiation",
    "version": "1.0.0"
  },
  "paths": {
...
    "/user/{user_id}/email": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Change Email",
        "description": "Request a change of the authenticated user's email address.\n\nSPEC-056 #3. This used to hand `{\"email\": ..., \"email_confirm\": True}` to the\nadmin API with no password check: the new address was marked verified on the\nspot, and neither the old nor the new inbox heard about it. A stolen access\ntoken was therefore a permanent account takeover \u2014 the attacker owns the\naddress that password resets go to, and the real owner has no notification\nand no way back.\n\nNow it costs the password, and the update runs on the user's OWN session, so\nSupabase treats it as a change request and mails the confirmation link. The\naddress does not move until someone clicks it.",
        "operationId": "change_email_user__user_id__email_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailUpdateSchema"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/users/{user_id}/profile": {
      "put": {
        "tags": [
          "System"
        ],
        "summary": "Update User Profile",
        "description": "Update a user's profile (display name, avatar).",
        "operationId": "update_user_profile_admin_users__user_id__profile_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfileUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}/avatar": {
      "post": {
        "tags": [
          "System"
        ],
        "summary": "Upload User Avatar",
        "description": "Upload a user avatar.",
        "operationId": "upload_user_avatar_admin_users__user_id__avatar_post",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_user_avatar_admin_users__user_id__avatar_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}/ban": {
      "put": {
        "tags": [
          "System"
        ],
        "summary": "Ban User",
        "description": "Ban or unban a user.",
        "operationId": "ban_user_admin_users__user_id__ban_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}/ai": {
      "get": {
        "tags": [
          "System"
        ],
        "summary": "Get User Ai Status",
        "description": "Get the current AI status for a user.",
        "operationId": "get_user_ai_status_admin_users__user_id__ai_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "System"
        ],
        "summary": "Toggle User Ai",
        "description": "Enable or disable AI for a specific user.",
        "operationId": "toggle_user_ai_admin_users__user_id__ai_put",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIToggleRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}": {
      "delete": {
        "tags": [
          "System"
        ],
        "summary": "Admin Delete User",
        "description": "Permanently delete a user: app-side data + the Supabase auth user.\nOrders are intentionally retained as business records.",
        "operationId": "admin_delete_user_admin_users__user_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "User Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIToggleRequest": {
        "properties": {
          "ai_enabled": {
            "type": "boolean",
            "title": "Ai Enabled"
          }
        },
        "type": "object",
        "required": [
          "ai_enabled"
        ],
        "title": "AIToggleRequest"
      },
...
      "BanRequest": {
        "properties": {
          "is_banned": {
            "type": "boolean",
            "title": "Is Banned"
          }
        },
        "type": "object",
        "required": [
          "is_banned"
        ],
        "title": "BanRequest"
      },
...
      "Body_upload_user_avatar_admin_users__user_id__avatar_post": {
        "properties": {
          "avatar": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "Avatar"
          }
        },
        "type": "object",
        "required": [
          "avatar"
        ],
        "title": "Body_upload_user_avatar_admin_users__user_id__avatar_post"
      },
...
      "EmailUpdateSchema": {
        "properties": {
          "new_email": {
            "type": "string",
            "title": "New Email"
          },
          "current_password": {
            "type": "string",
            "title": "Current Password"
          }
        },
        "type": "object",
        "required": [
          "new_email",
          "current_password"
        ],
        "title": "EmailUpdateSchema"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserProfileUpdateRequest": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          }
        },
        "type": "object",
        "title": "UserProfileUpdateRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}